This PR resolves a deprecation warning like below when the Ruby version is 2.7:

```
/path/to/ruby-kafka/spec/transaction_manager_spec.rb:594: warning: Passing the keyword argument as the last hash parameter is deprecated
/path/to/ruby-kafka/spec/transaction_manager_spec.rb:682: warning: The called method `txn_offset_commit_response' is defined here
```

## Before

```
% rspec --tag functional spec/transaction_manager_spec.rb:601 spec/functional/consumer_group_spec.rb:438
Run options:
  include {:functional=>true, :locations=>{"./spec/transaction_manager_spec.rb"=>[601], "./spec/functional/consumer_group_spec.rb"=>[438]}}
  exclude {:performance=>true, :fuzz=>true}

Kafka::TransactionManager
  #send_offsets_to_txn
    currently in transaction
/Users/arabiki/ghq/src/github.com/zendesk/ruby-kafka/spec/transaction_manager_spec.rb:594: warning: Passing the keyword argument as the last hash parameter is deprecated
/Users/arabiki/ghq/src/github.com/zendesk/ruby-kafka/spec/transaction_manager_spec.rb:682: warning: The called method `txn_offset_commit_response' is defined here
      notifies transaction coordinator

Consumer API
/Users/arabiki/ghq/src/github.com/zendesk/ruby-kafka/spec/spec_helper.rb:66: warning: Using the last argument as keyword parameters is deprecated; maybe ** should be added to the call
/Users/arabiki/ghq/src/github.com/zendesk/ruby-kafka/spec/spec_helper.rb:70: warning: The called method `create_topic' is defined here
  consuming messages with a custom assignment strategy

Finished in 3.72 seconds (files took 0.72687 seconds to load)
2 examples, 0 failures

```

## After

```
% rspec --tag functional spec/transaction_manager_spec.rb:601 spec/functional/consumer_group_spec.rb:438
Run options:
  include {:functional=>true, :locations=>{"./spec/transaction_manager_spec.rb"=>[601], "./spec/functional/consumer_group_spec.rb"=>[438]}}
  exclude {:performance=>true, :fuzz=>true}

Kafka::TransactionManager
  #send_offsets_to_txn
    currently in transaction
      notifies transaction coordinator

Consumer API
  consuming messages with a custom assignment strategy

Finished in 3.72 seconds (files took 0.61582 seconds to load)
2 examples, 0 failures

```